device/app/requirements.txt
backoff==1.11.1
certifi==2021.10.8
python-dotenv==0.19.2
ratelimit==2.2.1
urllib3==1.24.3
influxdb==5.3.1